+++ src/main/java/chess/SpringChessApplication.java
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SpringChessApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SpringChessApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(SpringChessApplication.class, args);
+    }
 }
