+++ src/main/java/chess/controller/GameController.java
@@ -1,36 +1,26 @@
 package chess.controller;
 
-import chess.dto.*;
+import chess.dto.ChessBoardDTO;
+import chess.dto.MoveDTO;
+import chess.dto.TurnDTO;
 import chess.service.ChessService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
-public class ChessController {
+public class GameController {
     private final ChessService chessService;
 
     @Autowired
-    public ChessController(ChessService chessService){
+    public GameController(ChessService chessService) {
         this.chessService = chessService;
-    }
-//
-    @GetMapping("/")
-    public String lobby() {
-        return "lobby";
     }
 
     @GetMapping("/{gameId}")
-    public String chess() {
+    public String initBoard() {
         return "index";
-    }
-
-    @GetMapping("/lobby/new")
-    @ResponseBody
-    public RoomIdDTO newGame(){
-        return chessService.newGame();
     }
 
     @GetMapping("/chessboard/{gameId}")
@@ -38,21 +28,18 @@ public class ChessController {
     public ChessBoardDTO loadGame(@PathVariable String gameId) {
         return chessService.loadGame(gameId);
     }
-//    gameid를 int로 받을지 string으로 받을지
+
+    //    gameid를 int로 받을지 string으로 받을지
     @GetMapping("/turn/{gameId}")
     @ResponseBody
     public TurnDTO turn(@PathVariable String gameId) {
         return chessService.turn(gameId);
     }
 
-    @PutMapping(path = "/move/{gameId}", consumes = "application/json")
-    public ResponseEntity move(@PathVariable String gameId, @RequestBody MoveDTO moveDTO){
-        System.out.println("@@@" + "shit");
-        System.out.println("@@@" + moveDTO.getSource());
+    @PutMapping(path = "/move/{gameId}")
+    public ResponseEntity move(@PathVariable String gameId, @RequestBody MoveDTO moveDTO) {
         return chessService.move(gameId, moveDTO);
-//        return null;
     }
-//
 //    @GetMapping
 //    public String {
 //        "/:id/result", ChessService::result}
