+++ src/main/java/chess/domain/piece/King.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.piece.movestrategy.NonIterableMoveStrategy;
+
 import java.util.List;
 
 public class King extends Piece {
