+++ src/main/java/chess/domain/piece/Knight.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.piece.movestrategy.NonIterableMoveStrategy;
+
 import java.util.List;
 
 public class Knight extends Piece {
