+++ src/main/java/chess/domain/piece/Pawn.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.piece.movestrategy.PawnMoveStrategy;
+
 import java.util.List;
 
 public class Pawn extends Piece {
