+++ src/main/java/chess/domain/piece/Rook.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.board.Direction;
 import chess.domain.piece.movestrategy.IterableMoveStrategy;
+
 import java.util.List;
 
 public class Rook extends Piece {
