+++ src/main/resources/logback-access.xml
@@ -5,5 +5,5 @@
         </encoder>
     </appender>
 
-    <appender-ref ref="STDOUT" />
+    <appender-ref ref="STDOUT"/>
 </configuration>
