+++ src/main/resources/templates/index.html
@@ -1,38 +1,38 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-  <meta charset="UTF-8">
-  <title>WOOWA CHESS</title>
-  <link rel="stylesheet" href="./css/index.css"/>
-  <link href="https://fonts.googleapis.com/css2?family=Jua&display=swap" rel="stylesheet">
-  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;700&display=swap" rel="stylesheet">
+    <meta charset="UTF-8">
+    <title>WOOWA CHESS</title>
+    <link rel="stylesheet" href="./css/index.css"/>
+    <link href="https://fonts.googleapis.com/css2?family=Jua&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;700&display=swap" rel="stylesheet">
 </head>
 <body>
 <div class="container">
-  <div class="title"><span>우아한 체스</span></div>
-  <div class="controller">
-    <button id="restart" class="control-btn">RESTART</button>
-    <button id="finish" class="control-btn">FINISH</button>
-    <button id="status" class="control-btn" disabled="true">STATUS</button>
-  </div>
-  <div class="content-wrapper">
-    <div class="chessBoard"></div>
-    <div class="status">
-      <div id="BLACK" class="player">
-        <img src="./images/player_lose.png" alt="player1"/>
-        <div class="score"></div>
-      </div>
-      <div class="b-turn">
-        <img class="black-turn" src="./images/up.png" alt="black-turn"/>
-        <img src="./images/chess_clock.png" alt="chess-clock"/>
-        <img class="white-turn" src="./images/down.png" alt="white-turn"/>
-      </div>
-      <div id="WHITE" class="player">
-        <img src="./images/player_lose.png" alt="player2"/>
-        <div class="score"></div>
-      </div>
+    <div class="title"><span>우아한 체스</span></div>
+    <div class="controller">
+        <button id="restart" class="control-btn">RESTART</button>
+        <button id="finish" class="control-btn">FINISH</button>
+        <button id="status" class="control-btn" disabled="true">STATUS</button>
     </div>
-  </div>
+    <div class="content-wrapper">
+        <div class="chessBoard"></div>
+        <div class="status">
+            <div id="BLACK" class="player">
+                <img src="./images/player_lose.png" alt="player1"/>
+                <div class="score"></div>
+            </div>
+            <div class="b-turn">
+                <img class="black-turn" src="./images/up.png" alt="black-turn"/>
+                <img src="./images/chess_clock.png" alt="chess-clock"/>
+                <img class="white-turn" src="./images/down.png" alt="white-turn"/>
+            </div>
+            <div id="WHITE" class="player">
+                <img src="./images/player_lose.png" alt="player2"/>
+                <div class="score"></div>
+            </div>
+        </div>
+    </div>
 </div>
 <script src="./js/lobby.js"></script>
 <script src="./js/main.js"></script>
