+++ src/main/resources/templates/lobby.html
@@ -1,19 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-  <meta charset="UTF-8">
-  <title>WOOWA CHESS</title>
-  <link rel="stylesheet" href="./css/index.css"/>
-  <link href="https://fonts.googleapis.com/css2?family=Jua&display=swap" rel="stylesheet">
-  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;700&display=swap" rel="stylesheet">
+    <meta charset="UTF-8">
+    <title>WOOWA CHESS</title>
+    <link rel="stylesheet" href="./css/index.css"/>
+    <link href="https://fonts.googleapis.com/css2?family=Jua&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;700&display=swap" rel="stylesheet">
 </head>
 <body>
 <div class="container">
-  <div class="title"><span>우아한 체스</span></div>
-  <div class="controller">
-    <button id="start" class="control-btn">START</button>
-    <button id="load" class="control-btn">LOAD</button>
-  </div>
+    <div class="title"><span>우아한 체스</span></div>
+    <div class="controller">
+        <button id="start" class="control-btn">START</button>
+        <button id="load" class="control-btn">LOAD</button>
+    </div>
 </div>
 <script src="./js/lobby.js"></script>
 </body>
